requirements/base.txt
# Storages
django-storages==1.9.1

# Boto3 (and dependencies)
boto3==1.12.48
botocore==1.15.48
docutils==0.15.2
jmespath==0.9.5
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
